Module Import.Logging

Logs tags attached to the log entries emitted by Irmin:

module Source_code_position : sig ... end

module type Clock = sig ... end

A monotonic time source. See Mtime_clock for an OS-dependent implementation.

val reporter : ?filter_src:( Logs.src -> bool ) -> ?prefix:string -> (module Clock) -> Logs.reporter

A default Logs reporter that is sensitive to the logs tags above.