Module 1-S.Commit

Commit defines immutable objects to describe store updates.

type t = commit

The type for store commits.

val t : repo -> t Irmin.Type.t

t is the value type for t.

val pp_hash : t Fmt.t

pp is the pretty-printer for commit. Display only the hash.

val v : repo -> info:info -> parents:commit_key list -> tree -> commit Lwt.t

v r i ~parents:p t is the commit c such that:

info c = i

parents c = p

tree c = t

val tree : commit -> tree

tree c is c's root tree.

val parents : commit -> commit_key list

parents c are c's parents.

val info : commit -> info

info c is c's info.

val hash : commit -> hash

hash c is c's hash.

Import/Export

val key : commit -> commit_key

key c is c's key.

val of_key : repo -> commit_key -> commit option Lwt.t

of_key r k is the the commit object in r with key k, or None if no such commit object exists.

val of_hash : repo -> hash -> commit option Lwt.t

of_hash r h is the commit object in r with hash h, or None if no such commit object is indexed in r.

Note: in stores for which commit_key = hash, this function has identical behaviour to of_key.